refactor: 주요 기능 및 함수를 클래스로 통합
