+++ src/main/kotlin/racingcar/game/Car.kt
@@ -1,3 +1,3 @@
-package racingcar
+package racingcar.game
 
 data class Car (val name: String, var position: Int = 0)
